psi: indexed element reads inherit the collection's arithmetic domain
place::analyze resolved self.coeffs[i] to the bare element type (i32/Exact)
while the 'in Wrapping'/'in Saturating' constraint lives on the array's
declaration -- so a domain-declared element read fed stores and temporaries
as if it were Exact. Under the strict typed-store rule that made the hoist
temp 'i32 in Wrapping = <i32 read>' an implicit-conversion rejection, and
mixed-domain diagnostics fired on the synthetic temps themselves. A direct
Indexed read whose element type declares no domain now inherits the
collection's domain; member projections keep the field's own type. Six
missed strictness migrations migrate with it (three field-counter invariants,
two composed-scalar programs, one lifetime-binder program, one transition
argument matrix, one pass fixture's exit_process argument) and the
saturating-element-guard fixture flips to checked.

+++ omega-rust/psi/pipeline/05_checked-trees-to-lowered-psi/src/tests/composed_scalar_results.rs
@@ -114,7 +114,7 @@ const LOOP_TO_EXPRESSION_LEAF: &str = r#"
             true -> count(n - 1, acc + 1)
             false -> leaf(acc)
         }
-        state leaf(&mut self, acc: i32 in Wrapping) -> i32 { acc + 100 }
+        state leaf(&mut self, acc: i32 in Wrapping) -> i32 { (acc + 100) as i32 }
     }
     machine Main::main(&mut self) -> i32 {
         let n: i32 = self.count(5, 0);
